+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Next
+
+* Improve performance of ASCII rendering - [@dblock](https://github.com/dblock).
+
 ## 1.0.0 (3/21/2014)
 
 * Initial public release - [@dblock](https://github.com/dblock).