+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## Next
 
 * Improve performance of ASCII rendering - [@dblock](https://github.com/dblock).
+* Render ASCII images on a background thread - [@dblock](https://github.com/dblock).
+* Use "Courier New" font instead of "Courier" - [@dblock](https://github.com/dblock).
 
 ## 1.0.0 (3/21/2014)