+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## Next
+## 1.1.0 (4/6/2014) 
 
 * [#2](https://github.com/dblock/ARASCIISwizzle/issues/2): Added [UIImage+ASCII](Classes/UIImage+ASCII.h) API - [@dblock](https://github.com/dblock).
 * Improved performance of ASCII rendering - [@dblock](https://github.com/dblock).